+++ keyboards/handwired/bongo80/rules.mk
@@ -1,4 +1,1 @@
-OLED_ENABLE = yes
-OLED_DRIVER_ENABLE = yes
-LTO_ENABLE = yes
-WPM_ENABLE = yes
+SRC += bongo80.c
